requirements/base.txt
Django>=1.4.5,<1.5

# Django apps
django-voting==0.1
South>=0.7.6,<0.8
django-assets>=0.8
django-taggit>=0.10.0

# Required by webassets, isn't shipped in Python < 2.7
argparse

# Libraries
cssmin>=0.1.4
lxml>=2.1.4

# For providing API
#django-tastypie

# Translate Toolkit
translate-toolkit>=1.10.0
# If you want to use Translate Toolkit 'master'
#-e git://github.com/translate/translate.git#egg=translate-toolkit
